requirements.txt
flask
flask-cors
tensorflow-cpu==2.15.0
numpy
pillow
